+++ collector/meminfo.go
@@ -1,0 +1,60 @@
+// Copyright 2015 The Prometheus Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// +build !nomeminfo
+
+package collector
+
+import (
+	"fmt"
+
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/common/log"
+)
+
+const (
+	memInfoSubsystem = "memory"
+)
+
+type meminfoCollector struct{}
+
+func init() {
+	Factories["meminfo"] = NewMeminfoCollector
+}
+
+// Takes a prometheus registry and returns a new Collector exposing
+// memory stats.
+func NewMeminfoCollector() (Collector, error) {
+	return &meminfoCollector{}, nil
+}
+
+// Update calls (*meminfoCollector).getMemInfo to get the platform specific
+// memory metrics.
+func (c *meminfoCollector) Update(ch chan<- prometheus.Metric) (err error) {
+	memInfo, err := c.getMemInfo()
+	if err != nil {
+		return fmt.Errorf("couldn't get meminfo: %s", err)
+	}
+	log.Debugf("Set node_mem: %#v", memInfo)
+	for k, v := range memInfo {
+		ch <- prometheus.MustNewConstMetric(
+			prometheus.NewDesc(
+				prometheus.BuildFQName(Namespace, memInfoSubsystem, k),
+				fmt.Sprintf("Memory information field %s.", k),
+				nil, nil,
+			),
+			prometheus.GaugeValue, v,
+		)
+	}
+	return nil
+}
